+++ package/lib/smooth_sheets.dart
@@ -7,7 +7,6 @@ export 'src/draggable/draggable_sheet.dart';
 export 'src/draggable/sheet_draggable.dart';
 export 'src/foundation/activities.dart';
 export 'src/foundation/animations.dart';
-export 'src/foundation/framework.dart';
 export 'src/foundation/keyboard_dismissible.dart';
 export 'src/foundation/notifications.dart';
 export 'src/foundation/physics.dart';
@@ -15,6 +14,7 @@ export 'src/foundation/sheet_content_scaffold.dart';
 export 'src/foundation/sheet_controller.dart'
     hide ImplicitSheetControllerScope, SheetControllerScope;
 export 'src/foundation/sheet_extent.dart';
+export 'src/foundation/sheet_viewport.dart';
 export 'src/foundation/theme.dart';
 export 'src/modal/cupertino.dart';
 export 'src/modal/modal_sheet.dart';
