+++ package/lib/src/foundation/sheet_container.dart
@@ -2,9 +2,9 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-import 'framework.dart';
 import 'sheet_controller.dart';
 import 'sheet_extent.dart';
+import 'sheet_viewport.dart';
 
 class SheetContainer extends StatelessWidget {
   const SheetContainer({
